Bugfix: Tried to delete an non existing token
When trying to confirm an email address with an invalid token it would render an error. It would also show a message saying the account has been successfully confirmed together with the message that the confirmation link is invalid or out-of-date.
I put everything in an else statement to prevent deleting non existing tokens or trying to log in a user with an invalid token.