Bugfix 2: Tried to delete an non existing token
When trying to change/update an email address with an invalid token it would render an error. It would also show a message saying the account has been successfully confirmed together with the message that the confirmation link is invalid.
Just like in the last pull request for this file (User.php) I put everything in an else statement to prevent deleting non existing tokens or trying to log in a user with an invalid token.