quicklogin feature when in DEBUG mode
Introduces a 'debug' module option. If set, a dropDown
with all users in the system is being displayed on the
login page. You can quickly login as every user without
a password.

YII_DEBUG and YII_ENV='dev' needs to be set in the application
index.php for this to work.