Revert "use Yii::->module->debug"
This reverts commit 807e510a3abf4ec0c48005b3891590da04b817df.

I thought about it and i disagree. The view file can be used outside of the user management module, for example when overriding the view file. In this case Yii::$app->module is null and we get an error. Referencing to the 'user' module explicitely seems to be the better option.