Added quicklogin feature (#794)
* quicklogin feature when in DEBUG mode

Introduces a 'debug' module option. If set, a dropDown
with all users in the system is being displayed on the
login page. You can quickly login as every user without
a password.

YII_DEBUG and YII_ENV='dev' needs to be set in the application
index.php for this to work.

* code style

* introduce helpers/Login helper to generate users list for quicklogin feature

* code style; use php short tags in security/login.php

* Ensure the module is not in DEBUG mode on production environments

* code style

* Improve YII_DEBUG & YII_ENV check

* use Yii::->module->debug

* Revert "use Yii::->module->debug"

This reverts commit 807e510a3abf4ec0c48005b3891590da04b817df.

I thought about it and i disagree. The view file can be used outside of the user management module, for example when overriding the view file. In this case Yii::$app->module is null and we get an error. Referencing to the 'user' module explicitely seems to be the better option.

* remove LoginHelper and move functionality to LoginForm

* introduce ensureCorrectDebugSetting() in Bootstrap.php

* fix CHANGELOG

* use ->module->debug

* code style

* code style